Split tilt.rb into several parts
